+++ statistics/BUILD.bazel
@@ -76,7 +76,7 @@ go_test(
     data = glob(["testdata/**"]),
     embed = [":statistics"],
     flaky = True,
-    shard_count = 41,
+    shard_count = 40,
     deps = [
         "//config",
         "//parser/ast",
